backend/requirements.txt
# FastAPI 和异步支持
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS 支持
python-dotenv==1.0.0

# QwenImg 核心依赖
dashscope>=1.14.0
Pillow>=9.0.0
requests>=2.25.0

# 工具
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
